requirements.txt
streamlit
pandas 
pinotdb
plotly.express 
plotly.graph_objects 
streamlit_autorefresh